refactoring PostItem
