+++ src/Components/PostItem/Modules/index.ts
@@ -1,0 +1,1 @@
+export * from "./PostHeader/PostHeader.Module";
