+++ src/Components/PostItem/PostItem.Component.tsx
@@ -5,37 +5,29 @@ import {
 } from "@expo/vector-icons";
 import ReadMore from "@fawazahmed/react-native-read-more";
 import { BottomSheetModal } from "@gorhom/bottom-sheet";
-import { MotiView } from "@motify/components";
-import { useAnimationState } from "@motify/core";
 import { useNavigation } from "@react-navigation/core";
 import { StackNavigationProp } from "@react-navigation/stack";
-import { formatRelative } from "date-fns";
-import { ptBR } from "date-fns/locale";
-import React, { useCallback, useMemo, useRef } from "react";
+import React, { useCallback, useRef } from "react";
 import { useWindowDimensions } from "react-native";
 import { TouchableWithoutFeedback } from "react-native-gesture-handler";
 import Toast from "react-native-toast-message";
-import CreatePost from "../CreatePost/CreatePost.Component";
 import { useUserContext } from "../../Contexts";
 import { BaseText } from "../../GlobalStyles/BaseStyles";
 import { BaseContainer } from "../../GlobalStyles/Containers.Style";
-import { Post, PostTypes } from "../../Interfaces/index";
+import { ClassRoomInterface, Post, PostTypes } from "../../Interfaces/index";
 import { MainStackParamList } from "../../Routes/MainStackNavigation/Interfaces";
 import { ROUTES_NAME } from "../../Routes/MainStackNavigation/RoutesName";
 import { baseApi, baseApiRoutes } from "../../Services";
-import { formatFilePathUrl } from "../../Utils";
 import ActivityItem from "../ActivityItem/ActivityItem.Component";
 import ConfirmationModal from "../ConfirmationModal/ConfirmationModal.Component";
-import ProfilePhoto from "../ProfilePhoto/ProfilePhoto.Component";
+import { PostHeader } from "./Modules";
 import {
   ActivityContainer,
   PostFooterContainer,
   PostItemContainer,
   styles,
   TextPostContainer,
-  TouchableHeader,
 } from "./Styles";
-import { ClassRoomInterface } from "../../Interfaces/index";
 
 interface PostItemProps {
   post: Post;
@@ -45,22 +37,6 @@ interface PostItemProps {
   classroom: ClassRoomInterface;
   isRenderedInPostDetailsPage?: boolean;
 }
-
-const IconWithTouchable: React.FC<{
-  onPress: () => void;
-  textTheme?: string;
-  iconName: string;
-  hideStyles?: boolean;
-}> = ({ onPress, textTheme, hideStyles = false, iconName }) => (
-  <TouchableWithoutFeedback onPress={onPress}>
-    <MaterialCommunityIcons
-      name={iconName as any}
-      size={25}
-      style={hideStyles ? undefined : styles.iconDots}
-      color={textTheme || "white"}
-    />
-  </TouchableWithoutFeedback>
-);
 
 /**
  * Post item listing component.
@@ -74,9 +50,9 @@ const PostItem: React.FC<PostItemProps> = ({
   isRenderedInPostDetailsPage = false,
   classroom,
 }) => {
-  const { height, width } = useWindowDimensions();
+  const { height } = useWindowDimensions();
   const navigation = useNavigation<StackNavigationProp<MainStackParamList>>();
-  const { user, userIsTeacher } = useUserContext();
+  const { user } = useUserContext();
   const goToPostDetails = useCallback(() => {
     navigation.navigate(ROUTES_NAME.POST_DETAILS, {
       post,
@@ -98,13 +74,6 @@ const PostItem: React.FC<PostItemProps> = ({
       routeIndexToReturnOnFinish: 2,
     });
   };
-  const postDate = useMemo(
-    () =>
-      formatRelative(new Date(post.createdAt), new Date(), {
-        locale: ptBR,
-      }),
-    [post.createdAt]
-  );
 
   const deletePost = useCallback(() => {
     return baseApi.delete(baseApiRoutes.POSTS + "/" + post._id).then((res) => {
@@ -115,27 +84,7 @@ const PostItem: React.FC<PostItemProps> = ({
     });
   }, []);
 
-  const dotsAnimationState = useAnimationState({
-    closed: {
-      height: 30,
-    },
-    open: {
-      height: 105,
-    },
-    from: {
-      height: 30,
-    },
-  });
-  const toggleDotsAnimation = () => {
-    dotsAnimationState.transitionTo(
-      dotsAnimationState.current === "closed" ||
-        dotsAnimationState.current === "from"
-        ? "open"
-        : "closed"
-    );
-  };
   const sheetRef = useRef<BottomSheetModal | null>(null);
-  const userIsAuthor = user?._id === post.author._id;
   return (
     <PostItemContainer deviceHeight={height}>
       <ConfirmationModal
@@ -143,78 +92,15 @@ const PostItem: React.FC<PostItemProps> = ({
         onConfirm={deletePost}
         modalRef={sheetRef}
       />
-      <BaseContainer
-        flexDirection="row"
-        align="center"
-        justify="space-between"
-        position="relative"
-        flex={1}
-      >
-        <TouchableHeader
-          deviceWidth={width}
-          onPress={goToPostDetails}
-          activeOpacity={0.1}
-          underlayColor="#dddd"
-          delayLongPress={500}
-        >
-          <>
-            <ProfilePhoto
-              size={50}
-              source={{
-                uri:
-                  formatFilePathUrl(post.author.profilePhoto?.path) ||
-                  "https://imgur.com/H5PWtBp.png",
-              }}
-            />
-            <BaseContainer width={`${width * 0.7}px`} flexDirection="column">
-              <BaseText
-                ellipsizeMode="tail"
-                numberOfLines={1}
-                marginLeft="13px"
-                color="black"
-              >
-                {post.author.name}
-              </BaseText>
-              <BaseText marginLeft="13px" fontSize="13px" color="grey">
-                {postDate}
-              </BaseText>
-            </BaseContainer>
-          </>
-        </TouchableHeader>
-        {(userIsTeacher || userIsAuthor) && (
-          <MotiView
-            style={{ ...styles.dotsContainer, backgroundColor: primaryTheme }}
-            state={dotsAnimationState}
-          >
-            <IconWithTouchable
-              onPress={toggleDotsAnimation}
-              iconName="dots-horizontal"
-              hideStyles
-              textTheme={textTheme}
-            />
-            {userIsAuthor && (
-              <CreatePost
-                classroom={classroom}
-                initialValues={post}
-                onPostCreation={getPosts as any}
-              >
-                {(ref) => (
-                  <IconWithTouchable
-                    onPress={() => ref.current?.present()}
-                    iconName="circle-edit-outline"
-                    textTheme={textTheme}
-                  />
-                )}
-              </CreatePost>
-            )}
-            <IconWithTouchable
-              iconName="delete"
-              textTheme={textTheme}
-              onPress={() => sheetRef.current?.present()}
-            />
-          </MotiView>
-        )}
-      </BaseContainer>
+      <PostHeader
+        goToPostDetails={goToPostDetails}
+        primaryTheme={primaryTheme}
+        textTheme={textTheme}
+        classroom={classroom}
+        getPosts={getPosts}
+        post={post}
+        sheetRef={sheetRef}
+      />
       <TextPostContainer>
         <ReadMore
           numberOfLines={isRenderedInPostDetailsPage ? 100 : 4}
